Changed double to Int

+++ Question3.java
@@ -13,7 +13,7 @@ public class Question3
     Scanner in = new Scanner(System.in); // Creating scanner object "in"
     
     System.out.print("");
-    double num = in.nextInt(); // Asssigning enterred user value into "num"
+    int num = in.nextInt(); // Asssigning enterred user value into "num"
 
     num = num * num;
